Move server app related code into app folder
