+++ packages/server/src/app/app.js
@@ -2,8 +2,8 @@ const express = require('express')
 const path = require('path')
 const cookieParser = require('cookie-parser')
 const logger = require('morgan')
-const indexRouter = require('./routes/index')
-const usersRouter = require('./routes/users')
+const indexRouter = require('./routes')
+const usersRouter = require('./models/user/user.controller')
 
 const app = express()
 
@@ -14,6 +14,6 @@ app.use(cookieParser())
 app.use(express.static(path.join(__dirname, 'public')))
 
 app.use('/', indexRouter)
-app.use('/users', usersRouter)
+app.use('/user', usersRouter)
 
 module.exports = app
