+++ packages/server/src/app/config/database-template.json
