+++ packages/server/src/app/models/user/user.controller.js
@@ -1,7 +1,6 @@
 const express = require('express')
 const router = express.Router()
 
-// GET users listing
 router.get('/', (_, res) => {
   res.send('respond with a resource')
 })
